Add missing verb to doctrine page
Previously the sentence (minus subordinate clauses) read "Yet it'd be hard that the power has doomed Ruby," which doesn't seem to be what's intended. Adding in "to argue that" seems to be the intent, so the core of the sentence is "Yet it'd be hard to argue that the power has doomed Ruby." I also added em-dashes to separate out a subordinate clause to make "it'd be hard to argue" easier to follow.